refactor(lint): Remove some egregious clang tidy suppressions (#160)
Closes #130 
Closes #146 
Closes #84 
Closes #152 
Closes #154 
Closes #137 
Closes #135 
Closes #139

+++ src/modules/engine/movement.cpp
@@ -10,7 +10,7 @@ void systemApplyVelocity(flecs::iter &it, size_t, const Velocity &v,
 }
 
 void systemUpdateExpiry(flecs::iter &it, size_t row, Expire &e) {
-  e.millis -= (it.delta_time() * 1000);
+  e.millis -= static_cast<int>(it.delta_time() * 1000);
   if (e.millis <= 0) {
     it.entity(row).destruct();
   }
